chore: moved existing filter logic to a separate hook
